refactor: Publish storage deals
Refactor publish storage deals to fail if any deals are invalid and report the error to the user
