+++ Cargo.toml
@@ -106,7 +106,7 @@ subxt = { version = "0.38.0" }
 subxt-signer = "0.38.0"
 syn = { version = "2.0.53" }
 tempfile = "3.10.1"
-thiserror = { version = "2.0.3" }
+thiserror = { version = "2.0.3", default-features = false }
 tokio = "1.37.0"
 tokio-stream = "0.1.15"
 tokio-util = "0.7.11"
