+++ cli/polka-storage/storagext/Cargo.toml
@@ -27,7 +27,7 @@ serde_json = { workspace = true }
 sha2 = { workspace = true }
 subxt = { workspace = true, features = ["jsonrpsee", "substrate-compat"] }
 subxt-signer = { workspace = true, features = ["subxt"] }
-thiserror = { workspace = true }
+thiserror = { workspace = true, default-features = true }
 tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
